requirements.txt
fastapi
uvicorn[standard]
pandas
RPi.GPIO
w1thermsensor

